split system mount point and reported mount point
We had the problem that freshly created volumes did not support
"automatic population". For this to work, the mounted directory has to
be empty. Ext4 for example creates a 'lost+found' in the freshly created
file system.

There is a reason that lost+found exists, so deleting it isn't the best
idea. It gets created by fsck, but what if the file system is that
broken that it can't create the dir/space for found files' meta data?
There is a reason mklost+found(8) exists.

To solve that, we simply mount the block device, to /something,
there is a /something/lost+found, but we simply create a
/something/data, which we expose as the "mountpoint". This allows us to:
- present a clean directory
- keep the lost+found for the worst case.

IMPORTANT: This breaks compat to the current plugin.
For migration you can copy the volumes data into a '/data' directory and
set the 'is-linstor-docker-volume' AUX property on the LINSTOR resource
definition.